requirements.txt
PyGObject
dbus-python
python-xlib
